IVS-547 - GRF007 - Valid vertical datum CRS type
