+++ features/steps/steps.py
@@ -1,4 +1,4 @@
 from givens import attributes, entities, relationships, values
 from thens import alignment, attributes, geometry, nesting, reference, relations, values, existence
 from steps import attribute_selection, attribute_value, entity_selection, model_traversal, representation, \
-    propertysets_and_qtys
+    propertysets_and_qtys, crs
